Adjust unit tests
