+++ src/ItemsStorage.php
@@ -230,25 +230,15 @@ final class ItemsStorage implements ItemsStorageInterface
     public function getParents(string $name): array
     {
         $rawItems = $this->getTreeTraversal()->getParentRows($name);
-        $items = [];
 
-        foreach ($rawItems as $rawItem) {
-            $items[$rawItem['name']] = $this->createItem(...$rawItem);
-        }
-
-        return $items;
+        return $this->getItemsIndexedByName($rawItems);
     }
 
     public function getChildren(string $name): array
     {
         $rawItems = $this->getTreeTraversal()->getChildrenRows($name);
-        $items = [];
 
-        foreach ($rawItems as $rawItem) {
-            $items[$rawItem['name']] = $this->createItem(...$rawItem);
-        }
-
-        return $items;
+        return $this->getItemsIndexedByName($rawItems);
     }
 
     public function hasChildren(string $name): bool
@@ -307,18 +297,15 @@ final class ItemsStorage implements ItemsStorageInterface
      */
     private function getItemsByType(string $type): array
     {
-        /** @psalm-var RawItem[] $rows */
-        $rows = $this
+        /** @psalm-var RawItem[] $rawItems */
+        $rawItems = $this
             ->database
             ->select()
             ->from($this->tableName)
             ->where(['type' => $type])
             ->fetchAll();
 
-        return array_map(
-            fn(array $row): Item => $this->createItem(...$row),
-            $rows,
-        );
+        return $this->getItemsIndexedByName($rawItems);
     }
 
     /**
@@ -479,5 +466,16 @@ final class ItemsStorage implements ItemsStorageInterface
         }
 
         return $this->treeTraversal;
+    }
+
+    private function getItemsIndexedByName(array $rawItems): array
+    {
+        $items = [];
+
+        foreach ($rawItems as $rawItem) {
+            $items[$rawItem['name']] = $this->createItem(...$rawItem);
+        }
+
+        return $items;
     }
 }
